Use of functions in parameter files, if possible
It would be interesting to mention how to use functions in parameter files, if possible, because I tried to use the [resourceGroup (). Location] "function, but without success. The same happened with the 'concat' function.
Sorry if it appear a dumb propose, but I'm a "new kid on the block" learning how to  deal and use ARM Templates.
Tks!